Change main_decks section to fixed height and add spacers to the left and right gutters.
This prevents the deck section from growing any larger than it needs to be. The
library will grow to consume all the vertical space. The spacers are for when
e.g. the mixer makes the main_decks section larger than the individual decks so
that the decks don't stretch vertically.